workers/metering-worker/requirements.txt
-e ../../libs/common
asyncpg>=0.29
structlog>=25.0
nats-py>=2.9.0
